Use numpy 64 bits (#65)
* Use numpy 64 bits

* fix error

+++ dataset/logic_circuit.py
@@ -12,70 +12,70 @@ import mnist
 def dset(d_name, num):
     if d_name == "or":
         dataset = list()
-        data = np.zeros((4 * num, 2), dtype=np.float128)
+        data = np.zeros((4 * num, 2))
         for i in range(num):
             data[4 * i] = [0, 0]
             data[4*i + 1] = [0, 1]
             data[4*i + 2] = [1, 0]
             data[4*i + 3] = [1, 1]
         dataset.append(data)
-        label = np.zeros((4 * num, 1), dtype=np.float128)
+        label = np.zeros((4 * num, 1))
         for i in range(4 * num):
             label[i] = data[i][0] or data[i][1]
         dataset.append(label)
         return dataset
     elif d_name == "and":
         dataset = list()
-        data = np.zeros((4 * num, 2), dtype=np.float128)
+        data = np.zeros((4 * num, 2))
         for i in range(num):
             data[4 * i] = [0, 0]
             data[4*i + 1] = [0, 1]
             data[4*i + 2] = [1, 0]
             data[4*i + 3] = [1, 1]
         dataset.append(data)
-        label = np.zeros((4 * num, 1), dtype=np.float128)
+        label = np.zeros((4 * num, 1))
         for i in range(4 * num):
             label[i] = data[i][0] and data[i][1]
         dataset.append(label)
         return dataset
     elif d_name == "nand":
         dataset = list()
-        data = np.zeros((4 * num, 2), dtype=np.float128)
+        data = np.zeros((4 * num, 2))
         for i in range(num):
             data[4 * i] = [0, 0]
             data[4*i + 1] = [0, 1]
             data[4*i + 2] = [1, 0]
             data[4*i + 3] = [1, 1]
         dataset.append(data)
-        label = np.zeros((4 * num, 1), dtype=np.float128)
+        label = np.zeros((4 * num, 1))
         for i in range(4 * num):
             label[i] = not (data[i][0] and data[i][1])
         dataset.append(label)
         return dataset
     elif d_name == "xor":
         dataset = list()
-        data = np.zeros((4 * num, 2), dtype=np.float128)
+        data = np.zeros((4 * num, 2))
         for i in range(num):
             data[4 * i] = [0, 0]
             data[4*i + 1] = [0, 1]
             data[4*i + 2] = [1, 0]
             data[4*i + 3] = [1, 1]
         dataset.append(data)
-        label = np.zeros((4 * num, 1), dtype=np.float128)
+        label = np.zeros((4 * num, 1))
         for i in range(4 * num):
             label[i] = data[i][0] ^ data[i][1]
         dataset.append(label)
         return dataset
     elif d_name == "w_not":
         dataset = list()
-        data = np.zeros((4 * num, 2), dtype=np.float128)
+        data = np.zeros((4 * num, 2))
         for i in range(num):
             data[4 * i] = [0, 0]
             data[4*i + 1] = [0, 1]
             data[4*i + 2] = [1, 0]
             data[4*i + 3] = [1, 1]
         dataset.append(data)
-        label = np.zeros((4 * num, 2), dtype=np.float128)
+        label = np.zeros((4 * num, 2))
         for i in range(num):
             label[4 * i] = [1, 1]
             label[4*i + 1] = [1, 0]
@@ -85,7 +85,7 @@ def dset(d_name, num):
         return dataset
     elif d_name == "mnist_train" or d_name == "mnist_test":
         (train_data, train_label), (test_data, test_label) = mnist.load_data()
-        data = np.zeros((num, 784 + 10), dtype=np.float128)
+        data = np.zeros((num, 784 + 10))
         for i in range(num):
             for j in range(28):
                 for k in range(28):
